fixlist  작업중
