+++ public/data/LIST_DATA/LOCATION.json
@@ -1,0 +1,132 @@
+{
+  "CITY": [
+    {
+      "id": 1,
+      "city": "서울",
+      "town": [
+        {
+          "id": "1",
+          "name": "강남 / 역삼 / 삼성"
+        },
+        {
+          "id": "2",
+          "name": "서초 / 교대 / 방배"
+        },
+        {
+          "id": "3",
+          "name": "사당"
+        },
+        {
+          "id": "4",
+          "name": "신사 / 압구정 / 청담"
+        },
+        {
+          "id": "5",
+          "name": "잠실 / 송파 "
+        },
+        {
+          "id": "6",
+          "name": "서울역 / 이태원 / 용산"
+        },
+        {
+          "id": "7",
+          "name": "명동 / 중구"
+        }
+      ]
+    },
+    {
+      "id": 2,
+      "city": "강원",
+      "town": [
+        {
+          "id": "1",
+          "name": "강릉 / 경포대 / 정동진"
+        },
+        {
+          "id": "2",
+          "name": "속초"
+        },
+        {
+          "id": "3",
+          "name": "양양"
+        },
+        {
+          "id": "4",
+          "name": "고성"
+        },
+        {
+          "id": "5",
+          "name": "동해 / 정선 "
+        },
+        {
+          "id": "6",
+          "name": "삼척"
+        },
+        {
+          "id": "7",
+          "name": "원주"
+        }
+      ]
+    },
+    {
+      "id": 3,
+      "city": "부산",
+      "town": [
+        {
+          "id": "1",
+          "name": "해운대"
+        },
+        {
+          "id": "2",
+          "name": "광안리 / 센텀"
+        },
+        {
+          "id": "3",
+          "name": "경성대"
+        },
+        {
+          "id": "4",
+          "name": "송정 / 기장"
+        },
+        {
+          "id": "5",
+          "name": "부산역 "
+        },
+        {
+          "id": "6",
+          "name": "동래"
+        },
+        {
+          "id": "7",
+          "name": "명지"
+        }
+      ]
+    },
+    {
+      "id": 4,
+      "city": "제주",
+      "town": [
+        {
+          "id": "1",
+          "name": "제주시"
+        },
+        {
+          "id": "2",
+          "name": "서귀포시 / 중문"
+        },
+        {
+          "id": "3",
+          "name": "제주국제공항"
+        },
+        {
+          "id": "4",
+          "name": "애월"
+        },
+        {
+          "id": "5",
+          "name": "성산"
+        }
+      ]
+    }
+  ]
+}
