+++ public/data/MyDaily/MYDAILY.json
@@ -1,0 +1,12 @@
+{
+  "mydaily": [
+    {
+      "id": 1,
+      "name": "윤탕헌",
+      "email": "changh950@naver.com",
+      "wishlist": "10",
+      "recentsaw": "15",
+      "reveiw": "5"
+    }
+  ]
+}
