+++ src/pages/RecentSaw/RecentSawNone.js
@@ -1,25 +1,30 @@
 import React from 'react';
+import { useHistory } from 'react-router';
 import styled from 'styled-components';
 
-function ReservationListNone() {
+function ResentSawNone() {
+  const history = useHistory();
+  const GotoMain = () => {
+    history.push(`/`);
+  };
+
   return (
     <ListBox>
-      <ReservationList>예약/구매내역이 없습니다.</ReservationList>
+      <ReservationList>최근 본 상품이 없습니다.</ReservationList>
       <ReservationDesk>
-        데일리호텔의 다양한 상품을 예약해보세요.
+        데일리호텔의 다양한 호텔을 구경해보세요.
       </ReservationDesk>
-      <GoMainButton>상품 둘러보기</GoMainButton>
+      <GoMainButton onClick={GotoMain}>상품 둘러보기</GoMainButton>
     </ListBox>
   );
 }
 
-export default ReservationListNone;
+export default ResentSawNone;
 
 const ReservationList = styled.div`
   font-size: 20px;
   font-weight: lighter;
   margin-bottom: 10px;
-  cursor: pointer;
 `;
 
 const ReservationDesk = styled.div`
@@ -34,6 +39,7 @@ const GoMainButton = styled.button`
   color: white;
   background-color: #6e2c9b;
   font-size: 17px;
+  cursor: pointer;
 `;
 
 const ListBox = styled.div`
